Service3/requirements.txt
boto3==1.35.33
botocore==1.35.33
mailersend==0.5.8
mysql_connector_python
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
